third_party/blink/web_tests/external/wpt/web-bundle/subresource-loading/script-relative-url.https.tentative.html

<!DOCTYPE html>
<title>Subresource loading using relative URLs in the 'resources' attribute</title>
<link rel="help" href="https://github.com/WICG/webpackage/blob/main/explainers/subresource-loading.md" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/test-helpers.js"></script>

<script>
  window.TEST_WEB_BUNDLE_ELEMENT_TYPE = 'script';
  setup(() => {
    assert_true(HTMLScriptElement.supports('webbundle'));
  });
</script>
<body>
  <script type="webbundle">
    {
      "source": "../resources/wbn/static-element.wbn",
      "resources": ["/web-bundle/resources/wbn/static-element/resources/script.js"]
    }
  </script>
  <script id="script" src="/web-bundle/resources/wbn/static-element/resources/script.js"></script>

  <script src="resources/relative-url-test.js"></script>
</body>
